requirements.txt
google-api-python-client==2.110.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-auth==2.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
flask==3.0.0
click==8.1.7
pyyaml==6.0.1
jinja2==3.1.2
markupsafe==2.1.3
html2text==2020.1.16
python-docx==1.1.0